[bt][at] Better optional field parsing
The AT command DSL allows specifying that certain argument fields are
optional if they are at the end of the sequence of arguments. It also
allows specifying that some arguments are variadic lists of possibly
optional elements.

This commit separates these two notions some more.  Previously an
optional argument would require a comma to delimit the missing
arguments, causing AT commands like "AT+1,".  However, this does not
match the single use we have for this feature, namely the +CLCC:
response.  This has optional arguments at the end, but they are not
comma delimited.

With this change, the delimiting comma is no longer present for optional
arguments but is still present for lists of optional items with the item
missing.

This change also updates the necessary fields in the +CLCC response
definition to be optional.

Test: fx test //src/connectivity/lib/at-commands:tests
Change-Id: I5f7d29359563482f60a3321598f5c25ff1de02e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327048
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>