[kernel] Clean up some stale physboot TODOs
The only meaningful code change is that we remove a 4GiB allocation
upper bound in physboot for x86, which was originally introduced for
back when start.S assumed that its page tables were within [0, 4GiB)
and physboot was handing off a root page table that it allocated. We
don't have this restriction any more.

Bug: 42164859
Change-Id: If4102e9578d0be8d6138a469433a910f84bea906
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1349526
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>