[fxtest] add --previous failed-tests flag
This CL adds the ability to print the failed tests from a previous test
run and the commands to run them again. The change builds on the
existing --previous flag by adding a new switch which will print out the
commands to run the previously failed tests.

Change-Id: I418ca188d669622304d5671226c68b4c9df857ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350053
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>