[gemini] Teach GEMINI.md about --previous failed-tests
This CL adds a line to the GEMINI.md file that instructs it to run
`fx test --previous failed-tests` to get a list of failed tests from the
previous run and how to test them.

TEST: run fx test with a failure and then run gemini with a prompt like
"why are my tests failing" and it will run the specified command to
figure out why it fails.

Change-Id: I28ee02acb9108a7d79daa9780f055777eeaebcb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350001
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Kyo Lee <kyol@google.com>