[driver_manager] Encapsulate node state
There are various states a node may be palced in, and not all of it's
members are valid in all states. Rather than place these members in
std::optionals, we can move that state into state specific structs,
accessed via a std::variant. As and additional step, we can centralize
control of state transitions more easily, although that is not handeled
in this CL.

There are a number of other cleanups also included here, such as setting
more methods to const or creating separate constructors for Node based
on whether it is a composite node or non-composite node.

Change-Id: I3b4b6c6f2ffaa0a50bbba6facd35c35595aa8bf0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1349668
Reviewed-by: Sarah Chan <spqchan@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>