[aml-gpu] Don't fail if optional compat protocol is not present
Rather than serve the compat protocol even when there is no metadata,
platform-bus is going to lazily serve it as necessary. By using service
validation, we can check whether or not it is served when attempting to
connect and fall back to the default at that point.

Change-Id: Ie6e3fc8b433cf1345c19a864bb2cdf67269b3f26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350324
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>