[aml-i2c] Don't fail if optional compat protocol isn't present
Rather than serve the compat protocol even when there is no metadata,
platform-bus is going to lazily serve it as necessary. By using service
validation, we can check whether or not it is served when attempting to
connect and fall back to the default at that point.

Change-Id: I8708a0ffb152edcba47551a379e8596cd4f11108
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350325
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Nick Eaton <nickeaton@google.com>