[netstack/filter] Take packet captures in tests
This will allow us to better debug failures in these tests; with all the
packet rewriting and filtering, it can be hard to pinpoint exactly where
things went wrong.

The total disk size of the packet captures for one run of
netstack-filter-fidl-tests is 3.6M, but most of the files are smaller
than the 4k sectors on my SSD. The raw size is 2.2M. As of writing,
there are 778 files.

A sampling of file names in the two formats that are output by this
change:

    forwarded_traffic_skips_local_ingress_ingress_v4_udp_router_client.pcapng
    forwarded_traffic_skips_local_ingress_ingress_v4_udp_router_server.pcapng
    forwarded_traffic_skips_local_ingress_ingress_v4_udp_src_port_default_router_client.pcapng
    forwarded_traffic_skips_local_ingress_ingress_v4_udp_src_port_default_router_server.pcapng
    forwarded_traffic_skips_local_ingress_ingress_v4_udp_src_port_inverse_match_router_client.pcapng
    forwarded_traffic_skips_local_ingress_ingress_v4_udp_src_port_inverse_match_router_server.pcapng
    drop_incoming_local_ingress_v4_all_traffic.pcapng
    drop_incoming_local_ingress_v4_dst_address_range_default.pcapng
    drop_incoming_local_ingress_v4_dst_address_range_inverse_match.pcapng
    drop_incoming_local_ingress_v4_dst_address_subnet_default.pcapng
    drop_incoming_local_ingress_v4_dst_address_subnet_inverse_match.pcapng

I spot-checked some of the files just to be sure and as far as I can
tell they're fine.

Bug: 439897891
Change-Id: I6a6a69644d89a7724066374fbeb5f5f6a7471145
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351204
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>