[ffx] clean up TargetQuery/TargetHandle matching
There were two paths (both being used!) for matching a query against a
handle: one was an explicit function (`query_matches_handle()`), the
other was an indirect path, converting the handle to a `Description`,
then matching the description against the query. The former was
actually incorrect, in that it did not correctly handle new address
types such as VSock, so it has been removed, and replaced with a method
on `TargetInfoQuery` that compares by going through the `Description`
conversion. (The code was also cleaned up to use the `From` trait.)

Along with the incorrect semantics, the previous query/handle matcher
did something odd: it returned _false_ for query::First. This is
because the function was only used in one way -- to determine whether
we should return early while waiting for discovery to return a
a target. If "First" _always_ matched, then we'd never report an
ambiguous target error. Therefore, it returned false.  To address this
while retaining the correct description-matching semantics, we simply
have an explicit check for First in the discovery-stream code.

Note that `Description` is primarily a daemon concept, as hinted at by
the fact that the conversion function ignores half the fields. While
the daemon is around, it makes sense to keep the data structure, but
eventually we could remove it, since non-daemon uses seem to be
restricted to the discovery crate, which means we could check against
the `TargetHandle` directly.

Test: Added unit tests for 'First' query resolution.

Change-Id: I6a6a6964e90399d1cddc764692e23534fc0a698d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1349364
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>