[reachability] Prevent updates after iface removal
It is possible for an interface removal event to be observed while there
is an in-progress Network Check. In this case, the interface state would
get updated to the result of the Network Check even if the state is
actually `Removed`. This leads to confusion especially when one of the
Internet pings succeeds before interface removal, as the interface state
is inaccurately reported as `Internet` after `Removed`.

Fixed: 413683922
Change-Id: I988b52eda8296f3689c6eee31741026c414bc155
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350052
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>