[trace] Adding SessionManager FIDL
This adds the SessionManager protocol for creating trace sessions.
It is a replacement for directly calling the Provisioner API which
This intent is to use the new API to start a trace session
that can continue to run after the host disconnects.

This CL does not include the implementation of the server
side of SessionManager, just  the client side.

Since the session_manager component is not available,
the trace functionality continues to utilize the daemon.

Bug: 431775370

Change-Id: Ifa8463d72d3f6c663dc306c9a83c2fcdcdf09a57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322612
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
API-Review: Clayton Wilkinson <wilkinsonclay@google.com>