[component] Fix broken Clone test
Remove unused "Empty" protocol from test helper, and replace usages of
EmptyCloneableNode with the Echo node. EmptyCloneableNode was being
erroneously used in tests as it did not verify correctly that cloning
succeeded, relying on a timeout for this purpose. In testing, it seems
that the server rarely processed the clone message, but the test was
still succeeding.

There was an equivalent test already defined for the Echo protocol, but
using the EchoString method to synchronize the client and server (thus
ensuring that the Clone call was processed). The remaining Clone tests
have been ported to use the Echo protocol instead, but otherwise remain
functional.

Note that there have been no changes to //sdk/lib/component itself, only
the tests.

Bug: 324080864
Test: fx test component_cpp_incoming_tests
Change-Id: Id91a991c656f52893dd26cc2e5983bb9d45669eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350246
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>