[sestarnix] Implement the audit_access/dontaudit pseudo-permission
When performing access-checks against an FsNode on behalf of a
task that invoked access() or similar, if any check (even a non-file
check such as for the "dac_override" capability permission) would
normally be audit-logged, logs may be suppressed via a `dontaudit`
statement for the `file { audit_access }` permission.

This is a re-land, the original landing of this CL having been
speculatively reverted.

Bug: 439840666
Change-Id: Ib39b4d101b0c2f0b2a8545fb9a3573dda15b1f74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351424
Reviewed-by: Marina Ciocea <marinaciocea@google.com>
Commit-Queue: Wez <wez@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>