[socket-proxy] Reduce logging for default network
When a change is made to the network registry, the current mark, DNS
servers, and default network is determined. Since all of these happen
together, there's no need to log each of them separately. We should log
it a single time when the default network is determined.

Change-Id: Ifc72fe805583861e09ad0014bf08f715d37b5e9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1341617
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>