[acpi] Fix duplicate composite duplicate names
Fix an issue with enumerating multiple composites that
share the same ACPI name, which results in duplicate
composite node spec names.

The ACPI bus now keep track of the ACPI composite names
and assign a unique ID for each composite, which is appended to
the composite node spec if it's greater than 0.

Bug: 419590561
Test: manual
Change-Id: I3727e361960936768201e6881f634c0a97a6252c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286425
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>