[trace][session_manager] Adding trace_session_manager component
This adds a component to trace_manager which manages the
lifecycle of a trace session. This allows the trace
session to be started, and then the host can disconnect and
reconnect to copy the trace data.

The primary goal of this feature is to allow capturing traces
when idle, so there is no usb or network connection to the host
which could preturb the trace results.

Bug: 431775370

Change-Id: Ie5394fefe2c63f04a1d344c1ab4b4885fe00db8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1344130
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Size-Review: Aidan Wolter <awolter@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>