[component_manager] fix race in storage admin integration test
The storage admin integration test would open an event stream to watch
for a stopped event of a child, but the child component was marked as
eager. This meant that in rare cases the child could start, do its work,
and stop before the test connects to the event stream, resulting in the
test missing the stop event.

Fixed: 440053040
Change-Id: I509aa8a848bad730378b0bbf1746e51ad8d6ce4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351065
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>