[kernel] Introduce ArchPostHandoffBootstrap()
We effectively want to move start.S logic to the very top of lk_main().
Currently the "first thing that must be called after hand-off" is called
HandoffFromPhys(). While we want to move the start.S logic to an
arch-specific subroutine of such a function, the current name is
arguably a weird fit since the logic in question isn't related to the
phys hand-off struct or phase per se. Accordingly, we rename
HandoffFromPhys() as PostHandoffBootstrap(), which retains the core
connotation of being sequenced immediately after hand-off, but also
allows for more general bootstrap logic to happen within. And we
introduce ArchPostHandoffBootstrap() as that arch-specific subroutine.

No logic is actually moved in this CL and the function is introduced as
a placeholder.

Bug: 42164859
Change-Id: I71191ff865812a3a1219028b1b70473ef741dc63
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350002
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>