[trace][perfetto] Map thread id to koid
This adds mapping thread ids from starnix containers to
Fuchsia koids. The result is Perfetto events are associated
with the correct thread tracks in the Perfetto UI.

The intention is to be able to use this information to handle mappings
when collecting traces from Perfetto via fuchsia_probes.

Bug: 433805146
Change-Id: I76906c845348d193d8405d4971744e43ae15fc2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1341746
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>