[kernel][vm] Mutex instead of Event in Evictor
The synchronization of multiple evictions was previously handled with
an event. This is no longer necessary and can be represented with a
Mutex instead. The benefit of a Mutex over an Event is that priority
can be propagated to the current holder of the Mutex, this prevents
priority inversion between high priority synchronous OOM eviction
from blocking behind low priority asynchronous eviction.

Bug: 423900604
Change-Id: I1a77b8f2af056643dcffa03db4b0a4ddd2a43248
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350704
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>