[libc][sanitizer] Temporarily suppress lsan-test
While we wait for the upstream lsan fix to land, temporarily disable
lsan-test on non-asan/hwasan targets. Keep it enabled when those are
provided so we still have some coverage. Note we disable the test in the
file itself because there isn't a good way to disable it on the build
level that allows the target to be built only on asan or hwasan builds.

Bug: 437346226
Change-Id: I6294866466c5681179af9d7e22592160f6ed336b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1349448
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>
Commit-Queue: Gulfem Savrun Yeniceri <gulfem@google.com>