[starnix] Add the ability to run Starnix tests against a Linux kernel.
This CL introduces the plumbing for Starnix test runner to make use of
the Machina virtualization environment. It modifies the existing
Starnix syscall tests to indicate a new test type, as well as a new
optional parameter to indicate the target kernel. Syscall tests that
target the Linux kernel now are ran against the Machina environment.

The ability to exercise these tests is currently gated behind adding
the virtualization framework to product assembly, as well as adding a
GN arg to build the Linux-based variant of tests. The GN arg is
temporary until we feel confident in the system.

Bug: 414635825
Test: fx test linux_syscalls_cpp_tests
Change-Id: I17e54b9889bbd9f8e0a506bc593e17cb3619b413
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1279520
Commit-Queue: Aaron Knobloch <aknobloch@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>