[third_party][fragile] Uprev to v2.0.1
v1.0.0 was yanked and has a bug where it would leak memory if the inner
type doesn't require drop to be called. This is because it stores it in
a Box and incorrectly doesn't drop that Box, leading to a leak.

Change-Id: I3ea3bf28fa8ffc5a020b4e618d9e0acb1b1d041c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1352056
Reviewed-by: David Koloski <dkoloski@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>