[assembly] Pass in release version info for product configs.
Previously, the ReleaseInfo.version field for a given product config was
copied over from the product settings struct. This worked, but in local
builds (where the version should be "unversioned") the version field was
being set to a timestamp, which included some characters that cannot be
used when communicating with the upstream versioning service.

This CL changes the product_config build process to use a passed-in
version string instead, like the other assembly artifacts. If it is
provided, use it. If not, use "unversioned".

Change-Id: I653bc687bb4d9e1a05412323cc2e61864baaf952
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345009
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>