[elf] Allow the UTC clock to be mappable.
This change allows the UTC clock to be mapped, which is necessary for
certain operations that require direct access to the system's
timekeeping mechanisms.

The `elf_runner` library is updated to support this mapping, ensuring
compatibility with systems that rely on mappable clocks.

Bug: 422533641
Change-Id: I04babf026247365e307988c5653f344dfc685bfe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351537
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>