[blobfs] use blob-exec when opening with exec
The CannotLoadBlobsExecutable opens a blob with Exec rights and then
checks if it can get an executable vmo from that file, which should
fail. The subtle behavior is that the open with Exec should _succeed_.
Do do that correctly we need to use the blob-exec capability which has
the exec right attached.

Change-Id: I5f280599299eee29edfecaa655cf453ae648fb6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1352584
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>