[elf] Put zx::Rights::INSPECT back into DUPLICATE_CLOCK_RIGHTS for UTC clock
Not having `INSPECT` is too restrictive when using mapped clocks, as it
prevents the caller from inspecting the clock's properties.

This change re-adds `INSPECT` to `DUPLICATE_CLOCK_RIGHTS` for UTC
clocks, allowing for proper inspection of mapped clocks.

While this is required for a distant change in code, it must be merged
first so that the elf runner for running the tests will have this code
in place when starting the tests.

Bug: 422533641
Change-Id: Iadbd372593343e79d4745c4858a014767a9868dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351030
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>