[starnix] Make waiter notifications signal-aware
The current waiter + notifications implementation is generic. For
signalling in particular, this presents a problem when multiple
resources (in this case signalfds) are polled together. A ready signal
for one signalfd ends up being broadcasted to all in the poll set,
regardless of the signalfd's mask of interested events.

Update the Waiter to accept a signal mask which can be used to filter
notifications that are not targeted to the Waiter.

Bug: 437146077
Test: fx test starnix-tests
Test: fx test -o starnix_syscalls_cpp_tests#meta/signalfd_test.cm

Change-Id: I9b850186bb8f86c2c9e5e94652aa625156f51504
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350127
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>