[scrutiny] Support reading delivery blobs instead of uncompressed blobs
This switches scrutiny to use the delivery_blob_type filed of
the product bundle's repository for what kind of blob to
expect, and teaches the FileArtifactReader how to properly
decompress delivery blobs.

It switches a number of direct File::open() calls to use
the ArtifactReader instead (so that the delivery blob
decompression will happen automatically).

Change-Id: Idf60c44f6baa75063d811ce01ba4b46217061d7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1352600
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>