[starnix] Ensure registers are only 32 bits when calling `apply_user_register`
The method signature is the same for aarch32 and aarch64. This CL mask
the bits of the register before and after calling the user provided
closure to ensure the register value is of the appropriate size.

Fixed: b/441172982
Change-Id: I0f119e7a53ae8d0c5f43d0de7ca1b0eddeab4c80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1354366
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>