[kernel][suspend] Fix resume timer wake source signaling/ack'ing
Fix an issue which showed up when testing on real HW with the newly
introduced wake source reporting.

Prior to the introduction of wake source reporting, acknowledging a
wake event object was an idempotent operation.  It was OK to ack a
wake even which had never been triggered.  With the addition of the
wake source reporting feature, this is no longer allowed.  There is
supposed to be a carefully followed protocol which ensures that by the
time a wake-event is acknowledged, it must be in the triggered state
and present on the `pending_list_`.

There exists a Very Special wake source in the system which represents
the "resume timer", which is responsible for waking the system after a
user specified deadline.  The wake-event for this wake source is
(logically) always acknowledged immediately after it becomes signaled.
Since the is no kernel Interrupt object for the user to ack, the
kernel just auto-acks the timer wake source.

Prior to this change, the wake event for this wake source became
signaled in the timer IRQ handler when the IRQ fired, and it was
*unconditionally ack'ed* every time the control thread for the suspend
operation unwound.

In the new wake source reporting world, this causes issues as it is
possible for a _different_ (non-resume timer) wake source to wake us
up, triggering a new bookkeeping assert as we attempt to ack a
wake-event which was never triggered.

Unfortunately, this is _very_ difficult to test except on real
hardware.  Non-timer interrupt which can wake the system vary on every
platform we have.  There is no "common hardware interrupt" we are
guaranteed to have which could be used to exercise this edge case.  It
may be possible to create a special QEMU instance that is guaranteed
to have a specific interrupt which we can cause to become triggered,
and use _that_ to write a test (which only runs on this special QEMU
configuration), but it will take some effort.

In the meantime, we need to fix the system as soon as we can.
Thankfully, the fix here is quite simple.

1) Don't simply "trigger" the resume timer wake event when its IRQ fires.
   Instead, use the dedicated WakeEvent::Strobe method which was created
   specifically to be used with the resume timer.
2) Don't unconditionally ack the timer wake-event when unwinding from suspend.
   We don't need to anymore.  Strobe takes care of the ack for us automatically
   now.

Fixed: 441330581
Change-Id: Ic4ea0e926082b88d64429fa4dd0f5a6b537df232
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1355364
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Commit-Queue: John Grossman <johngro@google.com>