[starnix][alarms] Plumb TargetTime to hr_alarm_manager.
No functionality is added or modified in this change. Plumbing the
change all the way will be the topic of a followup.

The old code was not using TargetTime. but used BootInstant instead to
specify wake alarm expiry. This change ensures that the TargetTime is
correctly plumbed to the hr_alarm_manager, which will allow us to
schedule wake alarms on the UTC timeline in addition to the boot
timeline.

Bug: 437984687
Change-Id: I18f8f69ec4f0dcbb4be43072a9b307fb62e1b80c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351228
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>