[kernel][vm] Hold LookupCursor owner lock
Moves the lock drop in the LookupCursor from InvalidateCursor to
EstablishCursor. This allows for callers to assume that when a page is
returned that lock of the owner of the page is held, at least up until
another page is requested.

Most usages of the LookupCursor do not need such a strong guarantee as
they can instead rely on the page being valid up until the leaf VMO
lock is taken. For example, PageFaultLocked contiguously holds the leaf
VMO lock, and to invalidate the page its lock would have to be acquired
to clear out any potentially page table mappings. Therefore allowing
PageFaultLocked to safely map the page, even if it does not hold the
owner lock.

However, although the page is validate to be mapped, the page metadata
in vm_page_t::object cannot be safely read or written without the owner
lock held. PromoteRangeForReclamation and ProtectRangeFromReclamation
attempt to do just this, and it mostly works since the cursor will not
always be invalidated while iterating and even if it is the owner will
not always be manipulating the page and its metadata. However, in an
unlikely race a kernel panic is possible.

Change-Id: I18bf8edff427ffe5f50b1a15673b72f1db8e29d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1354405
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>