[unwinder] Use named SP, LR, and PC from Starnix
These registers are explicitly placed by the Starnix CFI instructions in
restricted_return_loop, just in the context of 64 bit registers. We have
to convert these to the 32 bit equivalents so the ARM EHABI unwinder has
the values that it expects in the registers that it expects.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a69649749f971934507971cf915f1b8d145fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350286
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>