[ffx] Fix source-handling in discovery
libtarget was sometimes using `with_sources()` instead of
`set_sources()` which was a no-op -- it did not actually change the
sources used. This CL removes that method (except for use in tests),
and simplifies the resulting libtarget code, since the target-resolver
no longer needs to store a source.

Change-Id: I6a6a6964da8364e0c1c516fd08bbaa3ca02e79c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1352592
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>