[debugger] Detect discontiguous segment mappings
When using different loaders (such as in a linux environment within
starnix), it's possible for segments with large p_align values can get
mapped with enough memory between them such that another module is
entirely mapped between two segments of another module.

When this happens, we need to carefully track that the addresses match
up when we're calculating the end address of a particular module. When
they don't match is the signal that something was likely mapped in
between the segments.

Note that this will never happen with the Fuchsia loader.

Bug: 395606674
Change-Id: I6a6a696458b84755563b9cfb8c94577e6384ead5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351066
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>