[starnix] Rewrite __aeabi_uldivmod
This CL rewrites __aeabi_uldivmod so that it doesn't requires an adapter
asm method to use the correct calling convention. It uses uint128_t to
force the compiler to return the result in the registers r0 to r3.

Change-Id: I2ad5ef000ab75909591bed9747a339c087962500
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1355684
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>