[fvm] refactor resize tests to use new fvm instance
This will allow the tests to be parameterized over the fvm type so they
can be applied to the new fvm component as well as the driver. This
required an extensive refactor to the tests. Instead of rewriting the
GrowFvm free function that did a bunch of test logic, that logic is
brought directly into the test bodies, which make them longer but also
more clear what they are testing.

A new RestartFvmWithNewDiskSize method is added to the instance, which
lets the tests resize the test disk, which is core to what these tests
are testing.

The component version of these tests will be added as a followup.

Bug: 436889136
Change-Id: Ia2cac178c3130e2fbb1a0e23b346f169a8d79668
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350668
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>