[fvm] update pslice_count on startup if needed
It's possible that the fvm was provisioned with a size that doesn't
ultimately match the device it ends up on. In this case, it's expected
to expand up to the maximum amount it can allocate and update that
value.

This also updates the resize tests to work with the parameterized
structure, allowing it to target both fvm impls.

Bug: 436889136
Fixed: 440181963
Change-Id: I6993c0cec839cf9a282b23b3ec8e8967bab59695
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350669
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>