[clock] Remove the use of get_details() call temporarily.
This CL removes the use of `get_details()` call temporarily.

The `get_details()` call is being removed because it requires
`zx::Rights::INSPECT` that we're still adding to the UTC clock handle. We
also know that the clock size is one page, and this is checked in
downstream code, so the temporary impact should not be critical.

The intention is to put the call to `get_details` back once all
permissions are in place.

Bug: 415099415
Change-Id: I52398fc342746dc99afc1f2befa8df7ba47b33aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1356135
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>