[fidl][rust] Simpler transports and better errors
The division of work for proper error handling was not very clearly
defined between transports and FIDL clients and servers. This CL
simplifies the role of transports so that they no longer need to handle
closure, and moves that job to a new connection layer. It also ensures
that errors are correctly delayed and propagated according to the FIDL
specification.

Fixes: 440556959
Change-Id: Ia3931ecdf8c8e1a44e43b715a45ca29e38c3f3c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351844
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>