[fidl][rust] Improve connection closure ergonomics
Currently, connections do not drop the underlying transport until all
senders are dropped. This can cause straightforward code to unexpectedly
stall if the connection is closed, but the senders are waiting on client
or server termination to be dropped.

This change allows the underlying transport to be dropped as soon as the
connection is terminated. By adding a reference count to the connection
state, threads can detect the last use of the transport and drop it as
soon as it is no longer needed. This guarantees that as long as the
connection is stopped, the transport will be closed regardless of any
active futures. This fixes some testing situations which unexpectedly
stall, and make general use more robust.

Change-Id: Ie505e4593b402295babd12c200760484740c3bb0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1354844
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>