[starnix] Reduce copies of Credentials
current_creds() provides read access to a struct behind a lock, but
clones the Credentials instead of returning a reference. This
contributes to several percent of total cpu usage in benchmarks.

As this is often used to read just one or two fields of the credentials,
we waste a lot of time copying. Instead, we use a "with" function that
provides a reference to the Credentials so that we can copy only what we
need to.

Test: This change makes benchmarks happy.

Bug: 414642814
Change-Id: Ia0abe54659931bfdd66ed93cc7fa0c3424aeba77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1348766
Reviewed-by: Ambre Williams <ambre@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>