[kernel][x86] Delay reloading of GDT
We move the reloading of the GDT from start.S, which we're trying to
depopulate, back to the beginning of ArchPostHandoffBootstrap(). This
operation should be done as early as possible: ideally done before
touching the segment registers as that will cause the CPU to verify that
the GDTR points to valid memory for the GDT. Eventually, the GDT handed
off from physboot will be reallocated, but there's no danger of that in
early kernel bootstrap, and it's certainly fine to do in
ArchPostHandoffBootstrap(), which happens in the first handful of
instructions of lk_main().

Bug: 42164859
Change-Id: Ic7eee1806ae69e895107369a4b96ab2eb22e9e8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1355958
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>