[bazel][idk] Add idk_cc_source_library_zx() macro
`idk_cc_source_library_zx()` is a variant of `idk_cc_source_library()`
for libraries that are included in the IDK and used in the kernel. Such
libraries were previously defined in GN using the `zx_library()`
template and specifying `sdk_publishable`. bazel2gn uses `zx_library`
to define the GN target.

Support for kernel libraries not in the IDK will be added separately.

Due to the implied "include/" in the paths in `zx_library()`'s
`sdk_headers` parameter, it is easiest to add a new parameter that takes
the full path and can be used instead of `sdk_headers`. For consistency
with other templates, the new parameter name is `public`.

Bug: 429377203
Bug: 429374432

Change-Id: I802d3a82be39ab96924bec754f9f61d16e6a4c0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322608
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>