[assembly] Don't fallback to the product.build_info ...
... for release_info information.

This field will be the only option in some situations, but in-tree this
will often contain characters that are incompatible with the upstream
versioning server.

The product.build_info data is still used elsewhere in the fuchsia
build, but we should fallback to using "unversioned" if a version is not
explicitly provided for a given product configuration target.

Change-Id: I7b0e4db32adad620d000e4dbf312ece706b98806
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1357029
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>