[unwinder] Improve 32 bit frame transition conditions
Be more aggressive when asserting what registers are expected to be
available when attempting to transition to a 32 bit frame. We know that
as of today this transition should only ever succeed when we have
complete CFI directives provided by Starnix such that both PC and LR are
recovered and they should both be pointing to 32 bit code.

These expectations are now explicitly validated and an error returned if
they are not true.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a696448951f8a180a728259736348e644ac01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1355045
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>